Fix backslash bug
This PR aims to fix #11. I've also added tests, but still I'm not 100% sure that it won't break anything. I tried to make sure that the replacement phase titled "unescape double-quotes" does not mess with the escaped backslashes that are followed with a double quote.

Testing this with inline strings was so difficult for me, so created a folder named "tests" and put all the relevant files into that folder.

Closes #12